Consolidate packed consumer E2E journey
