+++ tests/e2e/browser/packaged-inline-html-theme.browser-e2e.spec.ts
@@ -5,15 +5,19 @@ import { existsSync, readFileSync } from "node:fs";
 import { mkdir, mkdtemp, readFile, readdir, rm, writeFile } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import { join, relative, resolve } from "node:path";
+import { pathToFileURL } from "node:url";
+import {
+	piPackedConsumerNpmEnv,
+	runPiPackedConsumerCommand,
+	runPiPackedConsumerNpm,
+} from "../test-utils/pi-packed-consumer-command.js";
 import {
 	capturePackagedCli,
-	cleanConsumerNpmEnv,
 	commandFailure,
 	createProjectAndSession,
 	getFreePort,
 	promptSession,
 	readToken,
-	runNpm,
 	startPackagedCli,
 	stopPackagedCli,
 	waitForHealth,
@@ -28,6 +32,23 @@ const PACKAGE_INSTALL_SEGMENTS = PACKAGE_NAME.split("/");
 const CANONICAL_BRIDGE_SIGNATURE = "data-bobbit-inline-theme-bridge";
 const SOURCE_BRIDGE_PATH = "src/shared/preview-bridge-scripts.ts";
 const THEME_TOKENS = ["--background", "--foreground", "--card", "--positive", "--chart-1"] as const;
+const PI_PACKAGES = [
+	"@earendil-works/pi-agent-core",
+	"@earendil-works/pi-ai",
+	"@earendil-works/pi-coding-agent",
+] as const;
+const INSPECTED_PACKAGES = [...PI_PACKAGES, "brace-expansion", "protobufjs"];
+const REQUIRED_PI_VERSION = "0.84.1";
+
+interface JsonRecord {
+	[key: string]: unknown;
+}
+
+interface DependencyOccurrence {
+	name: string;
+	version: string;
+	path: string[];
+}
 
 interface PackEntry {
 	name?: string;
@@ -48,21 +69,91 @@ interface ThemeState {
 
 interface RuntimeReport {
 	commands: CommandResult[];
+	pack?: unknown;
 	packFiles: string[];
+	selectedPiVersion?: string;
+	tree?: unknown;
+	binaries?: unknown;
 	bridgeAssets: string[];
 	requests: string[];
 	cliStdout?: string;
 	cliStderr?: string;
 }
 
-function parsePackResult(stdout: string): PackEntry {
-	let parsed: unknown;
-	try { parsed = JSON.parse(stdout); } catch (error) {
-		throw new Error(`npm pack emitted malformed JSON: ${(error as Error).message}\nstdout:\n${stdout}`);
+function asRecord(value: unknown, label: string): JsonRecord {
+	expect(value, `${label} must be an object`).not.toBeNull();
+	expect(typeof value, `${label} must be an object`).toBe("object");
+	expect(Array.isArray(value), `${label} must not be an array`).toBe(false);
+	return value as JsonRecord;
+}
+
+function parseJson(stdout: string, label: string): unknown {
+	expect(stdout.trim(), `${label} stdout must contain JSON`).not.toBe("");
+	try {
+		return JSON.parse(stdout);
+	} catch (error) {
+		throw new Error(`${label} emitted malformed JSON: ${(error as Error).message}\nstdout:\n${stdout}`, {
+			cause: error,
+		});
 	}
+}
+
+function parsePackResult(stdout: string): { entry: PackEntry; report: unknown } {
+	const parsed = parseJson(stdout, "npm pack");
 	expect(Array.isArray(parsed), "npm pack --json must return an array").toBe(true);
 	expect(parsed).toHaveLength(1);
-	return (parsed as PackEntry[])[0]!;
+	return { entry: (parsed as PackEntry[])[0]!, report: parsed };
+}
+
+function parseVersion(version: string, label = "version"): [number, number, number] {
+	const match = /^(\d+)\.(\d+)\.(\d+)$/.exec(version);
+	expect(match, `${label} must be an exact stable semver, received ${JSON.stringify(version)}`).not.toBeNull();
+	return [Number(match![1]), Number(match![2]), Number(match![3])];
+}
+
+function compareVersions(left: string, right: string): number {
+	const a = parseVersion(left, "selected version");
+	const b = parseVersion(right, "compatibility baseline");
+	for (let index = 0; index < a.length; index++) {
+		if (a[index] !== b[index]) return a[index]! - b[index]!;
+	}
+	return 0;
+}
+
+function collectDependencies(tree: unknown): DependencyOccurrence[] {
+	const occurrences: DependencyOccurrence[] = [];
+	const visit = (value: unknown, ancestors: string[]): void => {
+		const node = asRecord(value, `npm ls node at ${ancestors.join(" > ") || "root"}`);
+		const dependencies = node.dependencies;
+		if (dependencies === undefined) return;
+		const dependencyMap = asRecord(dependencies, `dependencies at ${ancestors.join(" > ") || "root"}`);
+		for (const [name, rawDependency] of Object.entries(dependencyMap)) {
+			const dependency = asRecord(rawDependency, `npm ls dependency ${name}`);
+			const path = [...ancestors, name];
+			if (INSPECTED_PACKAGES.includes(name as typeof INSPECTED_PACKAGES[number])) {
+				expect(typeof dependency.version, `${path.join(" > ")} must report a version`).toBe("string");
+				occurrences.push({ name, version: dependency.version as string, path });
+			}
+			visit(dependency, path);
+		}
+	};
+	visit(tree, []);
+	return occurrences;
+}
+
+function collectNpmProblems(tree: unknown): string[] {
+	const problems: string[] = [];
+	const visit = (value: unknown): void => {
+		const node = asRecord(value, "npm ls node");
+		if (node.problems !== undefined) {
+			expect(Array.isArray(node.problems), "npm ls problems must be an array").toBe(true);
+			for (const problem of node.problems as unknown[]) problems.push(String(problem));
+		}
+		if (node.dependencies === undefined) return;
+		for (const dependency of Object.values(asRecord(node.dependencies, "npm ls dependencies"))) visit(dependency);
+	};
+	visit(tree);
+	return problems;
 }
 
 async function listFiles(root: string): Promise<string[]> {
@@ -319,13 +410,14 @@ test.describe("packed Bobbit inline HTML runtime", () => {
 			// Browser-v2 global setup produces a content-addressed fresh dist first.
 			// npm pack therefore tests the same built artifact published to npx users
 			// without running an expensive build inside this E2E spec.
-			const packed = await runNpm(
+			const packed = await runPiPackedConsumerNpm(
 				["pack", "--json", "--ignore-scripts", "--pack-destination", packDir],
 				{ cwd: REPO_ROOT, timeoutMs: 3 * 60_000 },
 			);
 			report.commands.push(packed);
 			expect(packed.code, commandFailure(packed)).toBe(0);
-			const pack = parsePackResult(packed.stdout);
+			const { entry: pack, report: packReport } = parsePackResult(packed.stdout);
+			report.pack = packReport;
 			expect(pack.name).toBe(PACKAGE_NAME);
 			expect(typeof pack.filename).toBe("string");
 			report.packFiles = normalizedPackagePaths(pack);
@@ -341,20 +433,122 @@ test.describe("packed Bobbit inline HTML runtime", () => {
 				"tarball must contain compiled dist/ui CSS assets",
 			).toBe(true);
 
-			const tarballPath = join(packDir, pack.filename!);
+			const tarballPath = resolve(packDir, pack.filename!);
 			expect(existsSync(tarballPath), `npm pack did not create ${tarballPath}`).toBe(true);
-			const install = await runNpm(
-				["install", tarballPath, "--no-audit", "--no-fund", "--omit=optional"],
-				{ cwd: consumerDir, env: cleanConsumerNpmEnv(consumerDir), timeoutMs: 10 * 60_000 },
+			const consumerEnv = piPackedConsumerNpmEnv(consumerDir);
+			const lockConfig = await runPiPackedConsumerNpm(
+				["config", "get", "package-lock"],
+				{ cwd: consumerDir, env: consumerEnv, timeoutMs: 30_000 },
+			);
+			report.commands.push(lockConfig);
+			expect(lockConfig.code, commandFailure(lockConfig)).toBe(0);
+			expect(lockConfig.stdout.trim(), "clean consumer must use npm's normal package-lock=true default").toBe("true");
+
+			// CI prewarms every tarball selected by this clean consumer. Offline mode
+			// fails closed instead of letting this deterministic E2E consult a registry.
+			const install = await runPiPackedConsumerNpm(
+				["install", "--offline", tarballPath],
+				{ cwd: consumerDir, env: consumerEnv, timeoutMs: 10 * 60_000 },
 			);
 			report.commands.push(install);
 			expect(install.code, commandFailure(install)).toBe(0);
+			expect(existsSync(join(consumerDir, "package-lock.json")), "consumer install must create its own lockfile").toBe(true);
+			expect(
+				existsSync(join(
+					consumerDir,
+					"node_modules",
+					"@earendil-works",
+					"pi-coding-agent",
+					"npm-shrinkwrap.json",
+				)),
+				"published pi-coding-agent must include its dependency-owned shrinkwrap",
+			).toBe(true);
 
 			const installedRoot = join(consumerDir, "node_modules", ...PACKAGE_INSTALL_SEGMENTS);
 			const installedManifest = JSON.parse(await readFile(join(installedRoot, "package.json"), "utf8")) as {
 				bin?: Record<string, string>;
+				dependencies?: Record<string, string>;
 			};
 			expect(installedManifest.bin?.bobbit).toBe("dist/server/cli.js");
+			const piPins = PI_PACKAGES.map(name => installedManifest.dependencies?.[name]);
+			expect(piPins.every(pin => typeof pin === "string"), "packed Bobbit must declare all three Pi dependencies").toBe(true);
+			expect(new Set(piPins).size, "packed Bobbit must pin all three Pi packages to one version").toBe(1);
+			const selectedPiVersion = piPins[0]!;
+			parseVersion(selectedPiVersion, "selected Pi pin");
+			expect(selectedPiVersion, "packed Bobbit must pin Pi exactly to the supported version").toBe(REQUIRED_PI_VERSION);
+			report.selectedPiVersion = selectedPiVersion;
+
+			const lsResult = await runPiPackedConsumerNpm(
+				["ls", ...INSPECTED_PACKAGES, "--all", "--json"],
+				{ cwd: consumerDir, env: consumerEnv, timeoutMs: 2 * 60_000 },
+			);
+			report.commands.push(lsResult);
+			expect(lsResult.code, commandFailure(lsResult)).toBe(0);
+			const tree = parseJson(lsResult.stdout, "npm ls");
+			report.tree = tree;
+			expect(collectNpmProblems(tree), "npm ls must have no invalid, missing, stale, or extraneous edges").toEqual([]);
+			const occurrences = collectDependencies(tree);
+
+			for (const piPackage of PI_PACKAGES) {
+				const piOccurrences = occurrences.filter(entry => entry.name === piPackage);
+				expect(piOccurrences.length, `${piPackage} must appear in the packed consumer tree`).toBeGreaterThan(0);
+				expect(
+					[...new Set(piOccurrences.map(entry => entry.version))],
+					`${piPackage} must not have mixed or stale versions`,
+				).toEqual([selectedPiVersion]);
+				expect(
+					piOccurrences.every(entry => entry.path.includes(PACKAGE_NAME)),
+					`${piPackage} must resolve through the installed Bobbit package`,
+				).toBe(true);
+			}
+
+			const braceOccurrences = occurrences.filter(entry => entry.name === "brace-expansion");
+			expect(braceOccurrences.length, "brace-expansion must appear in the packed consumer tree").toBeGreaterThan(0);
+			expect(
+				braceOccurrences.every(entry => compareVersions(entry.version, "5.0.7") >= 0),
+				`every brace-expansion edge must be 5.0.7+: ${JSON.stringify(braceOccurrences)}`,
+			).toBe(true);
+
+			const protobufOccurrences = occurrences.filter(entry => entry.name === "protobufjs");
+			expect(protobufOccurrences.length, "protobufjs must appear in the packed consumer tree").toBeGreaterThan(0);
+			expect(
+				protobufOccurrences.every(entry => compareVersions(entry.version, "7.6.5") >= 0),
+				`Pi ${selectedPiVersion} must resolve every protobufjs edge to 7.6.5+: ${JSON.stringify(protobufOccurrences)}`,
+			).toBe(true);
+
+			const binariesModulePath = join(installedRoot, "dist", "server", "binaries.js");
+			const binaries = await import(pathToFileURL(binariesModulePath).href) as {
+				expectedBinaryPackage(): string | null;
+				getFdResolution(): { source: string; path: string | null; expectedPackage: string };
+				getRgResolution(): { source: string; path: string | null; expectedPackage: string };
+			};
+			const expectedBinaryPackage = binaries.expectedBinaryPackage();
+			const resolutions = {
+				fd: binaries.getFdResolution(),
+				rg: binaries.getRgResolution(),
+			};
+			report.binaries = { expectedBinaryPackage, resolutions };
+			if (expectedBinaryPackage) {
+				for (const [tool, resolution] of Object.entries(resolutions)) {
+					expect(resolution.source, `${tool} must resolve from ${expectedBinaryPackage}`).toBe("bundled");
+					expect(resolution.expectedPackage).toBe(expectedBinaryPackage);
+					expect(resolution.path, `${tool} bundled resolution must have a path`).not.toBeNull();
+					expect(existsSync(resolution.path!), `${tool} binary does not exist at ${resolution.path}`).toBe(true);
+					const smoke = await runPiPackedConsumerCommand(resolution.path!, ["--version"], {
+						cwd: consumerDir,
+						env: consumerEnv,
+						timeoutMs: 30_000,
+					});
+					report.commands.push(smoke);
+					expect(smoke.code, commandFailure(smoke)).toBe(0);
+					expect(`${smoke.stdout}\n${smoke.stderr}`.trim(), `${tool} --version must print its version`).not.toBe("");
+				}
+			} else {
+				testInfo.annotations.push({
+					type: "unsupported-binary-platform",
+					description: `${process.platform}-${process.arch} has no published Bobbit binary package`,
+				});
+			}
 			for (const required of [
 				join(installedRoot, "dist", "server", "cli.js"),
 				join(installedRoot, "dist", "ui", "index.html"),
@@ -505,6 +699,31 @@ test.describe("packed Bobbit inline HTML runtime", () => {
 			const switchedFrame = await iframeTheme(page);
 			expectThemeMatches(switchedFrame.current, switchedHost, "live-switched inline computed theme");
 			expect(switchedFrame.canonicalBridgeCount).toBe(1);
+			const palettePersist = await fetch(`${baseUrl}/api/preferences`, {
+				method: "PUT",
+				headers: preferenceHeaders,
+				body: JSON.stringify({ palette: "rose" }),
+			});
+			expect(
+				palettePersist.ok,
+				`failed to persist packaged rose palette: ${palettePersist.status} ${await palettePersist.clone().text()}`,
+			).toBe(true);
+
+			await page.reload({ waitUntil: "domcontentloaded" });
+			await expect(page.locator("body")).toHaveAttribute("data-shortcuts-ready", "1", { timeout: 30_000 });
+			const reloadedIframe = page.locator('iframe[title="theme-card.html"]');
+			await expect(reloadedIframe).toBeVisible({ timeout: 30_000 });
+			await expect.poll(
+				async () => (await iframeTheme(page)).capture?.background ?? "",
+				{ timeout: 20_000, message: "packaged inline bridge must execute after a full browser reload" },
+			).not.toBe("");
+			const reloadedHost = await hostTheme(page);
+			const reloadedFrame = await iframeTheme(page);
+			expect(reloadedHost.palette).toBe("rose");
+			expect(reloadedFrame.authoredScriptRan).toBe(true);
+			expect(reloadedFrame.canonicalBridgeCount).toBe(1);
+			expectThemeMatches(reloadedFrame.capture!, reloadedHost, "reloaded parse-time inline capture");
+			expectThemeMatches(reloadedFrame.current, reloadedHost, "reloaded inline computed theme");
 
 			const sourceRuntimeRequests = report.requests.filter(rawUrl => {
 				const url = new URL(rawUrl);
